Add clockBits function.
This function pumps tdi and collects tdo output with a tms termination
option based on an in/out arbitrary byte pointer and a count of bits. Does
not affect bits beyond final bit count(i.e. there can be some slack space
that is untouched...so initialization needs to be done outside carefully)